requirements.txt
telethon
python-dotenv
PySocks